refactor: 분산되어 있던 상수 분리
-chore: useGalleryHistorySave의 console.log 제거
