+++ client/src/@types/main.ts
@@ -1,8 +1,3 @@
-import springGround from "../assets/images/spring-ground.png";
-import summerGround from "../assets/images/summer-ground.png";
-import autumnGround from "../assets/images/autumn-ground.png";
-import winterGround from "../assets/images/winter-ground.png";
-import dreamGround from "../assets/images/dream-ground.png";
 export interface IPreviewGalleryData {
   userName: string;
   keywords: string[];
@@ -12,11 +7,4 @@ export interface IPreviewGalleryData {
 export type IMainDataResponse = {
   gallery: IPreviewGalleryData[];
   searchState: string;
-};
-export const GROUND_TEXTURES = {
-  DREAM: dreamGround,
-  SPRING: springGround,
-  SUMMER: summerGround,
-  AUTUMN: autumnGround,
-  WINTER: winterGround,
 };
