+++ client/src/MainPage/MainPage.tsx
@@ -9,14 +9,14 @@ import FullScreenModal from "../components/modal/FullScreenModal";
 import Footer from "../components/Footer";
 import { Toast } from "../components/Toast/Toast";
 import TOAST from "../components/Toast/ToastList";
+import Splash from "./components/SplashScreen/SplashScreen";
+import ThemeSeletor from "../components/ThemeSelector";
 
 import useError from "../hooks/useError";
 import userStore from "../store/user.store";
 import toastStore from "../store/toast.store";
 import mainStore from "../store/main.store";
 import FloatLayout from "../layouts/FloatLayout";
-import Splash from "./components/SplashScreen/SplashScreen";
-import ThemeSeletor from "../components/ThemeSelector";
 
 function CreateMonumentButton({ showModal }: { showModal: () => void }) {
   const isLoggedIn = userStore((store) => store.isLoggedIn);
