+++ client/src/MainPage/mapObjects/Ground.tsx
@@ -1,10 +1,12 @@
 import * as THREE from "three";
-import { useTexture } from "@react-three/drei";
-import galleryStore from "../../store/gallery.store";
-import { THEME } from "../../@types/gallery";
-import { GROUND_TEXTURES } from "../../@types/main";
-import { useFrame, useThree } from "@react-three/fiber";
 import { useRef } from "react";
+import { useFrame, useThree } from "@react-three/fiber";
+import { useTexture } from "@react-three/drei";
+
+import galleryStore from "../../store/gallery.store";
+import { THEME } from "../../constants/theme";
+import { GROUND_TEXTURES } from "../../constants/texture";
+import { MAIN_CAMERA_Z } from "../../constants/positions";
 
 export default function Ground() {
   const theme = galleryStore((store) => store.theme);
@@ -16,7 +18,7 @@ export default function Ground() {
     texture.offset.x = camera.position.x / 10;
     texture.offset.y = -camera.position.z / 10;
     groundRef.current.position.x = camera.position.x;
-    groundRef.current.position.z = camera.position.z - 20;
+    groundRef.current.position.z = camera.position.z - MAIN_CAMERA_Z;
   });
   return (
     <mesh receiveShadow ref={groundRef} position={[0, 0.1, 0]} rotation-x={-Math.PI / 2}>
