+++ client/src/constants/positions.ts
@@ -1,5 +1,5 @@
-{
-  "stone": [
+export const POSITION = {
+  stone: [
     [4, 0],
     [-4, 0],
     [0, 2],
@@ -27,9 +27,9 @@
     [2, -3],
     [-1, -3],
     [-2, -3],
-    [0, -4]
+    [0, -4],
   ],
-  "monument": [
+  monument: [
     [-20, -20],
     [-15, -20],
     [-10, -20],
@@ -73,6 +73,9 @@
     [5, 20],
     [10, 20],
     [15, 20],
-    [20, 20]
-  ]
-}
+    [20, 20],
+  ],
+};
+
+export const MAIN_CAMERA_Y = 10;
+export const MAIN_CAMERA_Z = 10;
