+++ client/src/hooks/useGalleryHistorySave.ts
@@ -9,13 +9,11 @@ export function useGalleryHistorySave() {
   function applyGallery(data: IGalleryMapData, userId: string, url: string | null = null) {
     setData(data, userId);
     setSourceUrl(data.theme);
-    console.log("applied", { data, userId });
     if (url !== null) window.history.pushState({ data, userId }, "", url);
   }
   function initializeGallery(data: IGalleryMapData, userId: string) {
     setData(data, userId);
     setSourceUrl(data.theme);
-    console.log("initialized", { data, userId });
     window.history.replaceState({ data, userId }, "");
   }
   return { applyGallery, initializeGallery };
